is running agent task in Agent status
