+++ src/app/core/_services/shared/bulk.service.ts
@@ -1,5 +1,5 @@
 import { catchError, finalize, map, mergeMap, toArray } from 'rxjs/operators';
-import { from, of, Observable, Subscription, forkJoin } from 'rxjs';
+import { Observable, Subscription, forkJoin, from, of } from 'rxjs';
 import { Injectable } from '@angular/core';
 
 import { GlobalService } from '../main.service';
@@ -8,10 +8,7 @@ import { GlobalService } from '../main.service';
   providedIn: 'root'
 })
 export class BulkService {
-
-  constructor(
-    private gs: GlobalService
-  ) {}
+  constructor(private gs: GlobalService) {}
 
   Items: any[];
   Value: any;
@@ -80,48 +77,44 @@ export class BulkService {
    * Displays a progress bar and one is complete return confirmation of action
    *
    * @param {number} percentage - Progress value
-  */
+   */
 
-    async performBulkUpdate(
-      progressCallback: (percentage: number) => void
-    ): Promise<boolean> {
-      const Items = this.Items;
-      const Value = this.Value;
-      const totalItems = Items.length;
-      let deletedItems = 0;
+  async performBulkUpdate(
+    progressCallback: (percentage: number) => void
+  ): Promise<boolean> {
+    const Items = this.Items;
+    const Value = this.Value;
+    const totalItems = Items.length;
+    let deletedItems = 0;
 
-      // Create an array to collect the results (true for success, false for failure)
-      const results: boolean[] = [];
+    // Create an array to collect the results (true for success, false for failure)
+    const results: boolean[] = [];
 
-      const itemObservables: Observable<boolean>[] = [];
+    const itemObservables: Observable<boolean>[] = [];
 
-      console.log(this.path)
+    Items.forEach((item) => {
+      const observable = this.gs.update(this.path, item, Value).pipe(
+        map(() => {
+          deletedItems++;
+          const progress = (deletedItems / totalItems) * 100;
+          progressCallback(progress);
+          return true; // Indicate success for each item
+        }),
+        catchError((error) => {
+          // Handle errors
+          return of(false); // Indicate failure for each item
+        })
+      );
 
-      Items.forEach((item) => {
-        const observable = this.gs.update(this.path, item, Value).pipe(
-          map(() => {
-            deletedItems++;
-            const progress = (deletedItems / totalItems) * 100;
-            progressCallback(progress);
-            return true; // Indicate success for each item
-          }),
-          catchError((error) => {
-            // Handle errors
-            return of(false); // Indicate failure for each item
-          })
-        );
+      itemObservables.push(observable);
+    });
 
-        itemObservables.push(observable);
-      });
+    // Use forkJoin to wait for all item observables to complete
+    const resultsArray = await forkJoin(itemObservables).toPromise();
 
-      // Use forkJoin to wait for all item observables to complete
-      const resultsArray = await forkJoin(itemObservables).toPromise();
+    // Check if any item deletion has failed
+    const hasFailure = resultsArray.some((result) => result === false);
 
-      // Check if any item deletion has failed
-      const hasFailure = resultsArray.some((result) => result === false);
-
-      return !hasFailure; // Return true if there are no failures
-    }
-
+    return !hasFailure; // Return true if there are no failures
+  }
 }
-
